src/test/java/QueryProcessorTest.java
import org.example.DBManager;
import org.example.QueryProcessor;
import org.example.TransactionManager;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import java.io.IOException;

import static org.junit.Assert.assertEquals;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.*;

@RunWith(MockitoJUnitRunner.class)
public class QueryProcessorTest {

    @Mock
    private DBManager dbManagerMock;

    @Mock
    private TransactionManager transactionManager;

    private QueryProcessor queryProcessor;

    @Before
    public void setUp() {
        queryProcessor = new QueryProcessor(dbManagerMock);
    }

//    @Test
//    public void testCreateDatabase_ValidQuery() {
//        when(dbManagerMock.createDatabase(anyString())).thenReturn("Database testdb created successfully.");
//        String result = queryProcessor.processQuery("CREATE DATABASE testdb");
//        assertEquals("Database testdb created successfully.", result);
//        verify(dbManagerMock).createDatabase("testdb".toUpperCase());
//    }
//
//    @Test
//    public void testCreateDatabase_InvalidQuery() {
//        String result = queryProcessor.processQuery("CREATE DATABASE");
//        assertEquals("Invalid query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testUseDatabase_ValidQuery() {
//        when(dbManagerMock.useDatabase(anyString())).thenReturn("Using database TESTDB.");
//        String result = queryProcessor.processQuery("USE DATABASE testdb");
//        assertEquals("Using database TESTDB.", result);
//        verify(dbManagerMock).useDatabase("testdb".toUpperCase());
//    }
//
//    @Test
//    public void testUseDatabase_InvalidQuery() {
//        String result = queryProcessor.processQuery("USE DATABASE");
//        assertEquals("Invalid query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testCreateTable_ValidQuery() {
//        when(dbManagerMock.createTable(anyString(), anyString())).thenReturn("Table users created successfully.");
//        String result = queryProcessor.processQuery("CREATE TABLE users (ID INT, NAME VARCHAR(50), AGE INT)");
//        assertEquals("Table users created successfully.", result);
//        verify(dbManagerMock).createTable("users".toUpperCase(), "ID INT, NAME VARCHAR(50), AGE INT");
//    }
//
//    @Test
//    public void testCreateTable_InvalidQuery() {
//        String result = queryProcessor.processQuery("CREATE TABLE");
//        assertEquals("Invalid CREATE TABLE query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
////    @Test
////    public void testInsertIntoTable_ValidQuery() {
////        when(dbManagerMock.insertIntoTable(anyString(), anyString())).thenReturn("1 row inserted into users.");
////        String result = queryProcessor.processQuery("INSERT INTO users VALUES (1, 'John Doe', 30)");
////        assertEquals("1 row inserted into users.", result);
////        verify(dbManagerMock).insertIntoTable("users".toUpperCase(), "1, 'John Doe', 30".toUpperCase());
////    }
//
//    @Test
//    public void testInsertIntoTable_InvalidQuery() throws IOException {
//        String result = queryProcessor.processQuery("INSERT INTO");
//        assertEquals("Invalid INSERT INTO query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testSelectFromTable_ValidQuery() throws IOException {
//        when(dbManagerMock.selectFromTable(anyString(), anyString())).thenReturn("ID\tNAME\tAGE\n1\tJohn Doe\t30\n2\tJane Smith\t25");
//        String result = queryProcessor.processQuery("SELECT FROM users WHERE AGE > 20");
//        assertEquals("ID\tNAME\tAGE\n1\tJohn Doe\t30\n2\tJane Smith\t25", result);
//        verify(dbManagerMock).selectFromTable("users".toUpperCase(), "AGE > 20");
//    }
//
//    @Test
//    public void testSelectFromTable_InvalidQuery() throws IOException {
//        String result = queryProcessor.processQuery("SELECT FROM");
//        assertEquals("Invalid SELECT FROM query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testSelectAllFromTable_InvalidQuery2() throws IOException {
//        String result = queryProcessor.processQuery("SELECT * FROM");
//        assertEquals("Invalid SELECT * FROM query syntax.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testSelectAllFromTable_ValidQuery() throws IOException {
//        when(dbManagerMock.selectAllFromTable(anyString())).thenReturn("ID\tNAME\tAGE\n1\tJohn Doe\t30\n2\tJane Smith\t25");
//        String result = queryProcessor.processQuery("SELECT * FROM users");
//        assertEquals("ID\tNAME\tAGE\n1\tJohn Doe\t30\n2\tJane Smith\t25", result);
//        verify(dbManagerMock).selectAllFromTable("users".toUpperCase());
//    }
//
//    @Test
//    public void testUpdateTable_ValidQuery() throws IOException {
//        when(dbManagerMock.updateTable(anyString(), anyString(), anyString())).thenReturn("2 records updated in users.");
//        String result = queryProcessor.processQuery("UPDATE users SET AGE=31 WHERE NAME='John Doe'");
//        assertEquals("2 records updated in users.", result);
//        verify(dbManagerMock).updateTable("users".toUpperCase(), "AGE=31", "NAME='John Doe'".toUpperCase());
//    }
//
//    @Test
//    public void testUpdateTable_InvalidQuery() {
//        String result = queryProcessor.processQuery("UPDATE users SET AGE=31");
//        assertEquals("Invalid UPDATE query syntax.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testDeleteFromTable_ValidQuery() {
//        when(dbManagerMock.deleteFromTable(anyString(), anyString())).thenReturn("1 record deleted from users.");
//        String result = queryProcessor.processQuery("DELETE FROM users WHERE NAME='John Doe'");
//        assertEquals("1 record deleted from users.", result);
//        verify(dbManagerMock).deleteFromTable("users".toUpperCase(), "NAME='John Doe'".toUpperCase());
//    }
//
//    @Test
//    public void testDeleteFromTable_InvalidQuery() {
//        String result = queryProcessor.processQuery("DELETE FROM");
//        assertEquals("Invalid DELETE FROM query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testDropTable_ValidQuery() {
//        when(dbManagerMock.dropTable(anyString())).thenReturn("Table users dropped successfully.");
//        String result = queryProcessor.processQuery("DROP TABLE users");
//        assertEquals("Table users dropped successfully.", result);
//        verify(dbManagerMock).dropTable("users".toUpperCase());
//    }
//
//    @Test
//    public void testDropTable_InvalidQuery() {
//        String result = queryProcessor.processQuery("DROP TABLE");
//        assertEquals("Invalid query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }
//
//    @Test
//    public void testInvalidQuery() {
//        String result = queryProcessor.processQuery("INVALID QUERY");
//        assertEquals("Invalid query.", result);
//        verifyNoInteractions(dbManagerMock);
//    }

    // Additional test cases for edge cases, exceptions, and additional functionalities

}
